flight_service/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
SQLAlchemy>=2.0,<3.0
alembic>=1.13,<2.0
psycopg[binary]>=3.2,<4.0
grpcio>=1.70,<2.0
grpcio-tools>=1.70,<2.0
grpcio-status>=1.70,<2.0
protobuf>=5.29,<6.0
googleapis-common-protos>=1.63,<2.0
redis>=5.2,<6.0
pydantic>=2.7,<3.0
pydantic-settings>=2.3,<3.0
python-dotenv>=1.0,<2.0
